requirements.txt
docopt==0.6.2
biopython==1.66
